requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
openpyxl==3.1.5
python-dotenv==1.0.1
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
opentelemetry-api>=1.20,<2.0
opentelemetry-sdk>=1.20,<2.0
opentelemetry-exporter-otlp-proto-grpc>=1.20,<2.0
opentelemetry-instrumentation-fastapi>=0.41b0
